+++ server/src/main/java/site/bannabe/server/domain/users/repository/BookmarkStationRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.lang.NonNull;
 import site.bannabe.server.domain.users.entity.BookmarkStations;
+import site.bannabe.server.domain.users.repository.querydsl.CustomBookmarkStationRepository;
 
 public interface BookmarkStationRepository extends JpaRepository<BookmarkStations, Long>, CustomBookmarkStationRepository {
